feat: separate framework
