+++ framework/entities/actionSpace.ts
