+++ framework/entities/card.ts
