+++ framework/entities/dice.ts
