+++ framework/entities/elementsBag.ts
