+++ framework/entities/game.ts
