+++ framework/entities/gameAction.ts
