+++ framework/entities/gameElement.ts
