+++ framework/entities/gameHistory.ts
