+++ framework/entities/gameState.ts
