+++ src/entities/nothanks/nothankscard.ts
@@ -1,4 +1,4 @@
-import Card from "../framework/card"
+import Card from "framework/entities/card"
 
 export class NoThanksCard extends Card {
     readonly value: number;
